Move BmpUtils to Utils dir
