+++ ArduinoSignController/src/Utils/BmpUtils.cpp
