+++ ArduinoSignController/src/Utils/BmpUtils.h
